refactor: rename & extract components
